+++ hindsight-control-plane/src/messages/de.json
@@ -534,7 +534,6 @@
     "allConsolidatedWithDate": "Alle Erinnerungen konsolidiert (zuletzt: {date})",
     "pendingConsolidation": "{count} Erinnerungen stehen zur Konsolidierung aus",
     "constellation": "Konstellation",
-    "graph": "Graph",
     "table": "Tabelle",
     "timeline": "Zeitleiste",
     "hidePanel": "Bereich ausblenden",
@@ -548,20 +547,6 @@
     "linkTypes": "Verknüpfungstypen",
     "nodes": "Knoten",
     "links": "Verknüpfungen",
-    "graphTitle": "Graph",
-    "linksWithCount": "Verknüpfungen ({count})",
-    "clickToFilter": "· klicken zum Filtern",
-    "semantic": "Semantisch",
-    "temporal": "Zeitlich",
-    "entity": "Entität",
-    "causal": "Kausal",
-    "displayTitle": "Anzeige",
-    "showLabels": "Beschriftungen anzeigen",
-    "performanceTitle": "Leistung",
-    "maxNodes": "Maximale Knoten",
-    "allLinksVisible": "Alle Verknüpfungen zwischen sichtbaren Knoten werden angezeigt.",
-    "limitedTo50Nodes": "⚠️ Aus Leistungsgründen auf 50 Knoten begrenzt. Gesamt: {count}",
-    "clickNodeForDetails": "Auf einen Knoten klicken, um Details anzuzeigen",
     "columnObservation": "Beobachtung",
     "columnMemory": "Erinnerung",
     "columnSources": "Quellen",
@@ -1452,15 +1437,6 @@
     "factTypeExperience": "Erfahrung",
     "factTypeObservation": "Beobachtung",
     "actionClearContent": "Inhalt löschen"
-  },
-  "graph2d": {
-    "loading": "Diagramm wird geladen...",
-    "emptyState": "Keine Erinnerungen zum Anzeigen",
-    "linkTypeCausal": "Kausal ({type})",
-    "linkTypeGeneric": "{type}-Verknüpfung",
-    "linkTooltipEntity": "Entität:",
-    "linkTooltipWeight": "Gewicht:",
-    "controlsHint": "Ziehen zum Verschieben • Scrollen zum Zoomen • Doppelklick auf Knoten zum Fokussieren • Klick auf den Hintergrund zum Zurücksetzen"
   },
   "constellation": {
     "instructions": "Scrollen zum Zoomen · Ziehen zum Verschieben · Hover zum Erkunden · Klicken zum Auswählen",
