+++ hindsight-control-plane/src/messages/en.json
@@ -534,7 +534,6 @@
     "allConsolidatedWithDate": "All memories consolidated (last: {date})",
     "pendingConsolidation": "{count} memories pending consolidation",
     "constellation": "Constellation",
-    "graph": "Graph",
     "table": "Table",
     "timeline": "Timeline",
     "hidePanel": "Hide panel",
@@ -548,20 +547,6 @@
     "linkTypes": "Link types",
     "nodes": "Nodes",
     "links": "Links",
-    "graphTitle": "Graph",
-    "linksWithCount": "Links ({count})",
-    "clickToFilter": "· click to filter",
-    "semantic": "Semantic",
-    "temporal": "Temporal",
-    "entity": "Entity",
-    "causal": "Causal",
-    "displayTitle": "Display",
-    "showLabels": "Show labels",
-    "performanceTitle": "Performance",
-    "maxNodes": "Max nodes",
-    "allLinksVisible": "All links between visible nodes are shown.",
-    "limitedTo50Nodes": "⚠️ Limited to 50 nodes for performance. Total: {count}",
-    "clickNodeForDetails": "Click a node to see details",
     "columnObservation": "Observation",
     "columnMemory": "Memory",
     "columnSources": "Sources",
@@ -1452,15 +1437,6 @@
     "factTypeExperience": "experience",
     "factTypeObservation": "observation",
     "actionClearContent": "Clear Content"
-  },
-  "graph2d": {
-    "loading": "Loading graph...",
-    "emptyState": "No memories to display",
-    "linkTypeCausal": "Causal ({type})",
-    "linkTypeGeneric": "{type} link",
-    "linkTooltipEntity": "Entity:",
-    "linkTooltipWeight": "Weight:",
-    "controlsHint": "Drag to pan • Scroll to zoom • Double-click node to focus • Click background to reset"
   },
   "constellation": {
     "instructions": "Scroll to zoom · Drag to pan · Hover to explore · Click to select",
