+++ hindsight-control-plane/src/messages/fr.json
@@ -534,7 +534,6 @@
     "allConsolidatedWithDate": "Tous les souvenirs consolidés (dernier : {date})",
     "pendingConsolidation": "{count} souvenirs en attente de consolidation",
     "constellation": "Constellation",
-    "graph": "Graphe",
     "table": "Tableau",
     "timeline": "Chronologie",
     "hidePanel": "Masquer le panneau",
@@ -548,20 +547,6 @@
     "linkTypes": "Types de liens",
     "nodes": "Nœuds",
     "links": "Liens",
-    "graphTitle": "Graphe",
-    "linksWithCount": "Liens ({count})",
-    "clickToFilter": "· cliquer pour filtrer",
-    "semantic": "Sémantique",
-    "temporal": "Temporel",
-    "entity": "Entité",
-    "causal": "Causal",
-    "displayTitle": "Affichage",
-    "showLabels": "Afficher les étiquettes",
-    "performanceTitle": "Performance",
-    "maxNodes": "Nœuds max",
-    "allLinksVisible": "Tous les liens entre les nœuds visibles sont affichés.",
-    "limitedTo50Nodes": "⚠️ Limité à 50 nœuds pour les performances. Total : {count}",
-    "clickNodeForDetails": "Cliquez sur un nœud pour voir les détails",
     "columnObservation": "Observation",
     "columnMemory": "Souvenir",
     "columnSources": "Sources",
@@ -1452,15 +1437,6 @@
     "factTypeExperience": "expérience",
     "factTypeObservation": "observation",
     "actionClearContent": "Effacer le contenu"
-  },
-  "graph2d": {
-    "loading": "Chargement du graphe...",
-    "emptyState": "Aucun souvenir à afficher",
-    "linkTypeCausal": "Causal ({type})",
-    "linkTypeGeneric": "Lien {type}",
-    "linkTooltipEntity": "Entité :",
-    "linkTooltipWeight": "Poids :",
-    "controlsHint": "Glisser pour déplacer • Défiler pour zoomer • Double-clic sur un nœud pour zoomer • Clic sur le fond pour réinitialiser"
   },
   "constellation": {
     "instructions": "Défiler pour zoomer · Glisser pour déplacer · Survoler pour explorer · Cliquer pour sélectionner",
