+++ hindsight-control-plane/src/messages/ja.json
@@ -534,7 +534,6 @@
     "allConsolidatedWithDate": "すべてのメモリが統合済み（最終：{date}）",
     "pendingConsolidation": "{count}件のメモリが統合待ち",
     "constellation": "コンステレーション",
-    "graph": "グラフ",
     "table": "テーブル",
     "timeline": "タイムライン",
     "hidePanel": "パネルを非表示",
@@ -548,20 +547,6 @@
     "linkTypes": "リンクの種類",
     "nodes": "ノード",
     "links": "リンク",
-    "graphTitle": "グラフ",
-    "linksWithCount": "リンク（{count}件）",
-    "clickToFilter": "・クリックでフィルター",
-    "semantic": "セマンティック",
-    "temporal": "時系列",
-    "entity": "エンティティ",
-    "causal": "因果",
-    "displayTitle": "表示",
-    "showLabels": "ラベルを表示",
-    "performanceTitle": "パフォーマンス",
-    "maxNodes": "最大ノード数",
-    "allLinksVisible": "表示中のノード間のすべてのリンクが表示されています。",
-    "limitedTo50Nodes": "⚠️ パフォーマンスのため50ノードに制限されています。合計：{count}",
-    "clickNodeForDetails": "ノードをクリックして詳細を確認",
     "columnObservation": "観察",
     "columnMemory": "メモリ",
     "columnSources": "ソース",
@@ -1452,15 +1437,6 @@
     "factTypeExperience": "経験",
     "factTypeObservation": "オブザベーション",
     "actionClearContent": "内容をクリア"
-  },
-  "graph2d": {
-    "loading": "グラフを読み込み中...",
-    "emptyState": "表示するメモリがありません",
-    "linkTypeCausal": "因果 ({type})",
-    "linkTypeGeneric": "{type} リンク",
-    "linkTooltipEntity": "エンティティ:",
-    "linkTooltipWeight": "ウェイト:",
-    "controlsHint": "ドラッグで移動 • スクロールでズーム • ノードをダブルクリックでフォーカス • 背景をクリックでリセット"
   },
   "constellation": {
     "instructions": "スクロールでズーム · ドラッグで移動 · ホバーで探索 · クリックで選択",
