+++ hindsight-control-plane/src/messages/ko.json
@@ -534,7 +534,6 @@
     "allConsolidatedWithDate": "모든 메모리 통합됨 (마지막: {date})",
     "pendingConsolidation": "{count}개 메모리 통합 대기 중",
     "constellation": "별자리",
-    "graph": "그래프",
     "table": "표",
     "timeline": "타임라인",
     "hidePanel": "패널 숨기기",
@@ -548,20 +547,6 @@
     "linkTypes": "링크 유형",
     "nodes": "노드",
     "links": "링크",
-    "graphTitle": "그래프",
-    "linksWithCount": "링크 ({count})",
-    "clickToFilter": "· 클릭하여 필터링",
-    "semantic": "의미적",
-    "temporal": "시간적",
-    "entity": "엔티티",
-    "causal": "인과적",
-    "displayTitle": "표시",
-    "showLabels": "레이블 표시",
-    "performanceTitle": "성능",
-    "maxNodes": "최대 노드",
-    "allLinksVisible": "표시된 노드 간의 모든 링크가 표시됩니다.",
-    "limitedTo50Nodes": "⚠️ 성능을 위해 50개 노드로 제한됩니다. 전체: {count}",
-    "clickNodeForDetails": "노드를 클릭하면 세부 정보를 볼 수 있습니다",
     "columnObservation": "관찰",
     "columnMemory": "메모리",
     "columnSources": "소스",
@@ -1452,15 +1437,6 @@
     "factTypeExperience": "경험",
     "factTypeObservation": "관찰",
     "actionClearContent": "콘텐츠 지우기"
-  },
-  "graph2d": {
-    "loading": "그래프 로딩 중...",
-    "emptyState": "표시할 메모리가 없습니다",
-    "linkTypeCausal": "인과 ({type})",
-    "linkTypeGeneric": "{type} 링크",
-    "linkTooltipEntity": "엔티티:",
-    "linkTooltipWeight": "가중치:",
-    "controlsHint": "드래그하여 이동 • 스크롤하여 확대/축소 • 노드 더블클릭으로 포커스 • 배경 클릭으로 초기화"
   },
   "constellation": {
     "instructions": "스크롤하여 확대/축소 · 드래그하여 이동 · 호버하여 탐색 · 클릭하여 선택",
