+++ hindsight-control-plane/src/messages/zh-CN.json
@@ -534,7 +534,6 @@
     "allConsolidatedWithDate": "所有记忆已整合（最近：{date}）",
     "pendingConsolidation": "{count} 条记忆待整合",
     "constellation": "星座图",
-    "graph": "图谱",
     "table": "表格",
     "timeline": "时间线",
     "hidePanel": "隐藏面板",
@@ -548,20 +547,6 @@
     "linkTypes": "链接类型",
     "nodes": "节点",
     "links": "链接",
-    "graphTitle": "图谱",
-    "linksWithCount": "链接（{count}）",
-    "clickToFilter": "· 点击筛选",
-    "semantic": "语义",
-    "temporal": "时间",
-    "entity": "实体",
-    "causal": "因果",
-    "displayTitle": "显示",
-    "showLabels": "显示标签",
-    "performanceTitle": "性能",
-    "maxNodes": "最大节点数",
-    "allLinksVisible": "所有可见节点之间的链接均已显示。",
-    "limitedTo50Nodes": "⚠️ 出于性能限制，最多显示 50 个节点。总计：{count}",
-    "clickNodeForDetails": "点击节点查看详情",
     "columnObservation": "观察",
     "columnMemory": "记忆",
     "columnSources": "来源",
@@ -1452,15 +1437,6 @@
     "factTypeExperience": "经验",
     "factTypeObservation": "观察",
     "actionClearContent": "清除内容"
-  },
-  "graph2d": {
-    "loading": "正在加载图谱...",
-    "emptyState": "暂无记忆可显示",
-    "linkTypeCausal": "因果 ({type})",
-    "linkTypeGeneric": "{type} 链接",
-    "linkTooltipEntity": "实体：",
-    "linkTooltipWeight": "权重：",
-    "controlsHint": "拖拽平移 • 滚动缩放 • 双击节点聚焦 • 点击背景重置"
   },
   "constellation": {
     "instructions": "滚动缩放 · 拖拽平移 · 悬停探索 · 点击选择",
