+++ hindsight-control-plane/src/messages/zh-TW.json
@@ -534,7 +534,6 @@
     "allConsolidatedWithDate": "所有記憶已整合（最近：{date}）",
     "pendingConsolidation": "{count} 條記憶待整合",
     "constellation": "星座圖",
-    "graph": "圖譜",
     "table": "表格",
     "timeline": "時間軸",
     "hidePanel": "隱藏面板",
@@ -548,20 +547,6 @@
     "linkTypes": "連結型別",
     "nodes": "節點",
     "links": "連結",
-    "graphTitle": "圖譜",
-    "linksWithCount": "連結（{count}）",
-    "clickToFilter": "· 點選篩選",
-    "semantic": "語義",
-    "temporal": "時間",
-    "entity": "實體",
-    "causal": "因果",
-    "displayTitle": "顯示",
-    "showLabels": "顯示標籤",
-    "performanceTitle": "效能",
-    "maxNodes": "最大節點數",
-    "allLinksVisible": "所有可見節點之間的連結均已顯示。",
-    "limitedTo50Nodes": "⚠️ 出於效能限制，最多顯示 50 個節點。總計：{count}",
-    "clickNodeForDetails": "點選節點檢視詳情",
     "columnObservation": "觀察",
     "columnMemory": "記憶",
     "columnSources": "來源",
@@ -1452,15 +1437,6 @@
     "factTypeExperience": "經驗",
     "factTypeObservation": "觀察",
     "actionClearContent": "清除內容"
-  },
-  "graph2d": {
-    "loading": "載入中圖譜...",
-    "emptyState": "尚無記憶可顯示",
-    "linkTypeCausal": "因果 ({type})",
-    "linkTypeGeneric": "{type} 連結",
-    "linkTooltipEntity": "實體：",
-    "linkTooltipWeight": "權重：",
-    "controlsHint": "拖曳平移 • 滾動縮放 • 雙擊節點聚焦 • 按一下背景重設"
   },
   "constellation": {
     "instructions": "滾動縮放 · 拖曳平移 · 懸停探索 · 按一下選取",
